% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacube.R
\name{as.datacube}
\alias{as.datacube}
\alias{is.datacube}
\alias{print.datacube}
\alias{summary.datacube}
\title{Structure a data frame to a datacube}
\usage{
as.datacube(x, ...)

is.datacube(x)

\method{print}{datacube}(x, ...)

summary.datacube(x, ...)
}
\arguments{
\item{x}{A data.frame.}

\item{tbl_df}{A tbl_df (tibble) object.}

\item{obs_id}{The observation identifier. Defaults to \code{NULL} in which case
the combination of the dimensions is used. If there are no dimensions present, then the
\code{row.names} will be used from \code{x}.}

\item{dim_names}{The columns in the data frame that are dimensions, and maybe used
for statistical aggregation.}

\item{measure_names}{The measurements in the data frame.}

\item{attribute_names}{Not measured attributes of the observations in the data.frame.}

\item{unit}{The unit of the measurements}
}
\value{
A datacube object, which has at least observation ids.
}
\description{
Coerce a data.frame or a tbl_df object into a datacube.
}
\details{
Loosely follows the \href{https://www.w3.org/TR/vocab-data-cube/}{The RDF Data Cube Vocabulary}.
See \href{https://www.w3.org/TR/vocab-data-cube/#dsd-dimensions}{Dimensions, attributes and measures}
of the Data Cube.
}
\examples{
as.datacube (x = iris[1:6,],
             obs_id = NULL,
             dim_names = NULL,
             measure_names = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),
             attribute_names = "Species"
             )
}
\concept{datacube functions}
