% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{attributes_dataset}
\alias{attributes_dataset}
\alias{attributes_add}
\title{Attributes of a dataset}
\usage{
attributes_dataset(x)

attributes_add(x, attributes, sdmx_attributes = NULL)
}
\arguments{
\item{x}{A data.frame or inherited tibble, data.frame, or a structured list.}

\item{attributes}{The name or column number of the attributes within the dataset.}

\item{sdmx_attributes}{The optional dimensions and attributes that conform with
SDMX. \code{c("time", "geo")} will mark the "time" and "geo" attributes as conforming to
sdmx. See \href{https://raw.githubusercontent.com/UKGovLD/publishing-statistical-data/master/specs/src/main/vocab/sdmx-attribute.ttl}{sdmx-attribute}.}
}
\description{
Attributes of a dataset
}
\details{
Do not confuse with \code{base::\link{attributes}}.
See the W3C and SDMX definition of a \href{https://www.w3.org/TR/vocab-data-cube/#dsd-dimensions}{attribute}.
}
