% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{attributes_measurements}
\alias{attributes_measurements}
\alias{attributes_measurements_add}
\title{Attributes of a dataset}
\usage{
attributes_measurements(x)

attributes_measurements_add(x, attributes, sdmx_attributes = NULL)
}
\arguments{
\item{x}{A data.frame or inherited tibble, data.frame, or a structured list.}

\item{attributes}{The name or column number of the attributes within the dataset.}

\item{sdmx_attributes}{The optional SDMX dimensions.}
}
\description{
Attributes of a dataset
}
\details{
Do not confuse with \code{base::\link{attributes}}, which applies to the attributes
of the entire dataset, and not each observation (measurement) row.
See the W3C and SDMX definition of a \href{https://www.w3.org/TR/vocab-data-cube/#dsd-dimensions}{attribute}.
}
