% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{attributes_measures}
\alias{attributes_measures}
\alias{attributes_measures<-}
\title{Attributes of a dataset}
\usage{
attributes_measures(x)

attributes_measures(x, sdmx_attributes = NULL) <- value
}
\arguments{
\item{x}{A data.frame or inherited tibble, data.frame, or a structured list.}

\item{sdmx_attributes}{The optional SDMX dimensions.}

\item{value}{The name or column number of the within the dataset.}
}
\value{
A data frame of the names, class, isDefinedBy, and codeList properties of the attributes columns of a
dataset following the datacube model.
}
\description{
Attributes of a dataset
}
\details{
Do not confuse with \code{base::\link{attributes}}, which applies to the attributes
of the entire dataset, and not each observation (measurement) row.
See the W3C and SDMX definition of a \href{https://www.w3.org/TR/vocab-data-cube/#dsd-dimensions}{attribute}.
}
