% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avocado.R
\docType{data}
\name{avocado}
\alias{avocado}
\title{Avocado Dataset}
\format{A data frame with 18249 rows and 13 variables:
\describe{
\item{\code{Date}}{double. DESCRIPTION.}
\item{\code{AveragePrice}}{double. DESCRIPTION.}
\item{\code{Total Volume}}{double. DESCRIPTION.}
\item{\code{4046}}{double. DESCRIPTION.}
\item{\code{4225}}{double. DESCRIPTION.}
\item{\code{4770}}{double. DESCRIPTION.}
\item{\code{Total Bags}}{double. DESCRIPTION.}
\item{\code{Small Bags}}{double. DESCRIPTION.}
\item{\code{Large Bags}}{double. DESCRIPTION.}
\item{\code{XLarge Bags}}{double. DESCRIPTION.}
\item{\code{type}}{character. DESCRIPTION.}
\item{\code{year}}{double. DESCRIPTION.}
\item{\code{region}}{character. DESCRIPTION.}
}}
\usage{
avocado
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
