% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defined.R
\name{c.haven_labelled_defined}
\alias{c.haven_labelled_defined}
\title{Combine Values into a defined Vector}
\usage{
\method{c}{haven_labelled_defined}(...)
}
\arguments{
\item{...}{objects to be concatenated.}
}
\value{
A haven_labelled_defined vector.
}
\description{
The c() method with the haven_labelled_defined class requires a strict
matching of the var_label, unit, definiton, and namespace attributes (if
they exist and do not have a \code{NULL} value)
}
\examples{
a <- defined(1:3, label="Length", unit="meter")
b <- defined(4:6, label="Length", unit="meter")
c(a,b)
}
\seealso{
\code{\link[=defined]{defined()}}
}
