% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churn.R
\docType{data}
\name{churn}
\alias{churn}
\title{Churn_Modelling Dataset}
\format{A data frame with 10000 rows and 14 variables:
\describe{
\item{\code{RowNumber}}{double. DESCRIPTION.}
\item{\code{CustomerId}}{double. DESCRIPTION.}
\item{\code{Surname}}{character. DESCRIPTION.}
\item{\code{CreditScore}}{double. DESCRIPTION.}
\item{\code{Geography}}{character. DESCRIPTION.}
\item{\code{Gender}}{character. DESCRIPTION.}
\item{\code{Age}}{double. DESCRIPTION.}
\item{\code{Tenure}}{double. DESCRIPTION.}
\item{\code{Balance}}{double. DESCRIPTION.}
\item{\code{NumOfProducts}}{double. DESCRIPTION.}
\item{\code{HasCrCard}}{double. DESCRIPTION.}
\item{\code{IsActiveMember}}{double. DESCRIPTION.}
\item{\code{EstimatedSalary}}{double. DESCRIPTION.}
\item{\code{Exited}}{double. DESCRIPTION.}
}}
\usage{
churn
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
