% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit.R
\docType{data}
\name{credit_train}
\alias{credit_train}
\title{GiveMeSomeCredit training Dataset}
\format{A data frame with 150000 rows and 11 variables:
\describe{
\item{\code{SeriousDlqin2yrs}}{double. DESCRIPTION.}
\item{\code{RevolvingUtilizationOfUnsecuredLines}}{double. DESCRIPTION.}
\item{\code{age}}{double. DESCRIPTION.}
\item{\code{NumberOfTime30-59DaysPastDueNotWorse}}{double. DESCRIPTION.}
\item{\code{DebtRatio}}{double. DESCRIPTION.}
\item{\code{MonthlyIncome}}{character. DESCRIPTION.}
\item{\code{NumberOfOpenCreditLinesAndLoans}}{double. DESCRIPTION.}
\item{\code{NumberOfTimes90DaysLate}}{double. DESCRIPTION.}
\item{\code{NumberRealEstateLoansOrLines}}{double. DESCRIPTION.}
\item{\code{NumberOfTime60-89DaysPastDueNotWorse}}{double. DESCRIPTION.}
\item{\code{NumberOfDependents}}{character. DESCRIPTION.}
}}
\usage{
credit_train
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
