\name{dds}
\alias{dds}
\docType{data}
\title{
Data sample for the Dataset package
}
\description{
A small data base for testing purpose only. These data are under license, you can't use it in your work. Thank you for your comprehension.

The \code{dds} name stands for 'Dataset Data Sample'.
}
\usage{data(dds)}
\format{
A Dataset object with 15 rows and 10 columns.
}
\details{
Run the \code{exportPDF} function on the database for getting details.
}
\source{
NA
}
\references{
NA
}
\examples{
## data(dds)
## exportPDF(dds, description.chlength = 60)
}
\keyword{datasets}
\keyword{data}
\keyword{Dataset}
\keyword{dds}
