% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-package.R
\docType{package}
\name{dataset-package}
\alias{dataset}
\alias{dataset-package}
\title{dataset: Sample Dataset in LG Leadership Academy}
\description{
Sample Dataset for R education.
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

Authors:
\itemize{
  \item Youngwoo Kim \email{stats7445@gmail.com} [contributor]
  \item SeonghoNa NA \email{drkevin2022@gmail.com} [contributor]
}

}
\keyword{internal}
