% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\alias{is.dataset}
\title{Structure a data frame to dataset}
\usage{
dataset(x, dimensions, measures, attributes, ...)

is.dataset(x)
}
\arguments{
\item{x}{A data.frame or interited tibble, data.frame, or a structured list.}

\item{dimensions}{The name or column number of the dimensions within the dataset.}

\item{measures}{The name or column number of the measures within the dataset.}

\item{attributes}{The name or column number of the attributes within the dataset.}

\item{...}{sdmx_attributes}
}
\description{
A constructor for the dataset class.
}
\details{
Loosely follows the \href{https://www.w3.org/TR/vocab-data-cube/}{The RDF Data Cube Vocabulary},
but with far stricter definition.
}
