% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{dataset_bibentry}
\alias{dataset_bibentry}
\title{Create a bibentry for a well-described data frame}
\usage{
dataset_bibentry(dataset)
}
\arguments{
\item{dataset}{A dataset.}
}
\value{
A bibentry.
}
\description{
This is a wrapper around \code{\link[utils:bibentry]{bibentry}} to correctly
use the attributes added to a dataset by \code{\link{datacite}}.
}
\examples{
datacite_iris <- datacite_dataset(
   df = iris,
   Title = "Iris Dataset",
   Creator = person("Anderson", "Edgar", role = "aut"),
   Publisher=" American Iris Society",
   PublicationYear = 1935,
   Geolocation = "US",
   Language = "en")

dataset_bibentry(datacite_iris)
}
\seealso{
datacite
}
\concept{citation functions}
