% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describe.dataset}
\title{Describe a dataset object}
\usage{
describe(x)

\method{describe}{dataset}(x)
}
\arguments{
\item{x}{A dataset object.}
}
\value{
No object is returned, but they key attributes are printed on the screen.
}
\description{
Describe a dataset object
}
