% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dimensions}
\alias{dimensions}
\alias{dimensions_add}
\title{Dimensions of a dataset}
\usage{
dimensions(x)

dimensions_add(x, dimensions, sdmx_attributes = NULL)
}
\arguments{
\item{x}{A data.frame or interited tibble, data.frame, or a structured list.}

\item{dimensions}{The name or column number of the dimensions within the dataset.}

\item{sdmx_attributes}{The optional dimensions and attributes that conform with
SDMX. \code{c("time", "geo")} will mark the "time" and "geo" attributes as conforming to
sdmx. See \href{https://raw.githubusercontent.com/UKGovLD/publishing-statistical-data/master/specs/src/main/vocab/sdmx-attribute.ttl}{sdmx-attribute}.}
}
\description{
Dimensions of a dataset
}
\details{
Do not confuse with \code{base::\link{dim}}. The \href{https://www.w3.org/TR/vocab-data-cube/#dsd-dimensions}{dimension} in the definition
of the DataSet is different from the 'dimension' definition of the R language.
}
\examples{
df <- data.frame ( sex = c("M", "F"), value = c(1,2))
dimensions_add(df, "sex", sdmx_attributes = "sex")
}
\seealso{
\code{\link{base::dim}}
}
