% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dublincore.R
\name{dublincore_to_triples}
\alias{dublincore_to_triples}
\title{Convert Dublin Core list to RDF triples}
\usage{
dublincore_to_triples(dclist, dataset_id = "http://example.com/dataset")
}
\arguments{
\item{dclist}{A named list of Dublin Core properties and values}

\item{dataset_id}{The URI subject of the dataset}
}
\value{
A character vector of RDF triples in N-Triples format
}
\description{
Convert Dublin Core list to RDF triples
}
\keyword{internal}
