% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp.R
\docType{data}
\name{emp_attr}
\alias{emp_attr}
\title{Employee_Attrition_Performance Dataset}
\format{A data frame with 1470 rows and 35 variables:
\describe{
\item{\code{Age}}{double. DESCRIPTION.}
\item{\code{Attrition}}{character. DESCRIPTION.}
\item{\code{BusinessTravel}}{character. DESCRIPTION.}
\item{\code{DailyRate}}{double. DESCRIPTION.}
\item{\code{Department}}{character. DESCRIPTION.}
\item{\code{DistanceFromHome}}{double. DESCRIPTION.}
\item{\code{Education}}{double. DESCRIPTION.}
\item{\code{EducationField}}{character. DESCRIPTION.}
\item{\code{EmployeeCount}}{double. DESCRIPTION.}
\item{\code{EmployeeNumber}}{double. DESCRIPTION.}
\item{\code{EnvironmentSatisfaction}}{double. DESCRIPTION.}
\item{\code{Gender}}{character. DESCRIPTION.}
\item{\code{HourlyRate}}{double. DESCRIPTION.}
\item{\code{JobInvolvement}}{double. DESCRIPTION.}
\item{\code{JobLevel}}{double. DESCRIPTION.}
\item{\code{JobRole}}{character. DESCRIPTION.}
\item{\code{JobSatisfaction}}{double. DESCRIPTION.}
\item{\code{MaritalStatus}}{character. DESCRIPTION.}
\item{\code{MonthlyIncome}}{double. DESCRIPTION.}
\item{\code{MonthlyRate}}{double. DESCRIPTION.}
\item{\code{NumCompaniesWorked}}{double. DESCRIPTION.}
\item{\code{Over18}}{character. DESCRIPTION.}
\item{\code{OverTime}}{character. DESCRIPTION.}
\item{\code{PercentSalaryHike}}{double. DESCRIPTION.}
\item{\code{PerformanceRating}}{double. DESCRIPTION.}
\item{\code{RelationshipSatisfaction}}{double. DESCRIPTION.}
\item{\code{StandardHours}}{double. DESCRIPTION.}
\item{\code{StockOptionLevel}}{double. DESCRIPTION.}
\item{\code{TotalWorkingYears}}{double. DESCRIPTION.}
\item{\code{TrainingTimesLastYear}}{double. DESCRIPTION.}
\item{\code{WorkLifeBalance}}{double. DESCRIPTION.}
\item{\code{YearsAtCompany}}{double. DESCRIPTION.}
\item{\code{YearsInCurrentRole}}{double. DESCRIPTION.}
\item{\code{YearsSinceLastPromotion}}{double. DESCRIPTION.}
\item{\code{YearsWithCurrManager}}{double. DESCRIPTION.}
}}
\usage{
emp_attr
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
