% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp.R
\docType{data}
\name{emp_tmnt}
\alias{emp_tmnt}
\title{Employee_10Year_Termination Dataset}
\format{A data frame with 49653 rows and 18 variables:
\describe{
\item{\code{EmployeeID}}{double. DESCRIPTION.}
\item{\code{recorddate_key}}{character. DESCRIPTION.}
\item{\code{birthdate_key}}{character. DESCRIPTION.}
\item{\code{orighiredate_key}}{character. DESCRIPTION.}
\item{\code{terminationdate_key}}{character. DESCRIPTION.}
\item{\code{age}}{double. DESCRIPTION.}
\item{\code{length_of_service}}{double. DESCRIPTION.}
\item{\code{city_name}}{character. DESCRIPTION.}
\item{\code{department_name}}{character. DESCRIPTION.}
\item{\code{job_title}}{character. DESCRIPTION.}
\item{\code{store_name}}{double. DESCRIPTION.}
\item{\code{gender_short}}{character. DESCRIPTION.}
\item{\code{gender_full}}{character. DESCRIPTION.}
\item{\code{termreason_desc}}{character. DESCRIPTION.}
\item{\code{termtype_desc}}{character. DESCRIPTION.}
\item{\code{STATUS_YEAR}}{double. DESCRIPTION.}
\item{\code{STATUS}}{character. DESCRIPTION.}
\item{\code{BUSINESS_UNIT}}{character. DESCRIPTION.}
}}
\usage{
emp_tmnt
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
