% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\docType{data}
\name{games}
\alias{games}
\title{Video Game Sales Dataset}
\format{A data frame with 16598 rows and 11 variables:
\describe{
\item{\code{Rank}}{double. DESCRIPTION.}
\item{\code{Name}}{character. DESCRIPTION.}
\item{\code{Platform}}{character. DESCRIPTION.}
\item{\code{Year}}{character. DESCRIPTION.}
\item{\code{Genre}}{character. DESCRIPTION.}
\item{\code{Publisher}}{character. DESCRIPTION.}
\item{\code{NA_Sales}}{double. DESCRIPTION.}
\item{\code{EU_Sales}}{double. DESCRIPTION.}
\item{\code{JP_Sales}}{double. DESCRIPTION.}
\item{\code{Other_Sales}}{double. DESCRIPTION.}
\item{\code{Global_Sales}}{double. DESCRIPTION.}
}}
\usage{
games
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
