% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codelist.R
\name{get_codelist}
\alias{get_codelist}
\title{Get a codelist}
\usage{
get_codelist(codelist = "CL_FREQ", var_name = "id")
}
\arguments{
\item{codelist}{The abbreviation of the codelist.}

\item{var_name}{The name of the \code{id} identifier variable that is coded.}
}
\value{
A codelist in a data frame.
}
\description{
Get a codelist
}
