% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prefix.R
\name{get_prefix}
\alias{get_prefix}
\alias{get_resource_identifier}
\title{Get prefix/resource identifier from CURIE}
\usage{
get_prefix(curie)

get_resource_identifier(curie)
}
\arguments{
\item{curie}{A Compact URI, for example, \code{'eg:refArea'}.}
}
\value{
The prefix, for example, \code{'eg'} from \code{'eg:refArea'}.
}
\description{
Separate the \code{eg}: prefix and the \code{refArea} resource identifier
from the Compact URI \code{`eg:refArea`} or \code{`[eg:refArea]`}
}
\examples{
get_prefix("eg:refArea")
get_resource_identifier("eg:refArea")

}
