% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelabel.R
\name{is.prelabelled}
\alias{is.prelabelled}
\title{Test if a vector is prelabelled}
\usage{
is.prelabelled(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Logical scalar.
}
\description{
Determine whether an object inherits from the
\code{"prelabelled"} class.
}
