% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_unit.R
\name{labelled_unit}
\alias{labelled_unit}
\alias{is.labelled_unit}
\alias{as.character.haven_labelled_unit}
\alias{summary.haven_labelled_unit}
\title{Create a labelled vector with defined unit of measure}
\usage{
labelled_unit(x = double(), labels = NULL, label = NULL, unit = NULL)

is.labelled_unit(x)

\method{as.character}{haven_labelled_unit}(x, ...)

\method{summary}{haven_labelled_unit}(object, ...)
}
\arguments{
\item{x}{A vector to label. Must be either numeric (integer or double) or
character.}

\item{labels}{A named vector or \code{NULL}. The vector should be the same type
as \code{x}. Unlike factors, labels don't need to be exhaustive: only a fraction
of the values might be labelled.}

\item{label}{A short, human-readable description of the vector.}

\item{unit}{The unit of measure described with a character string of length one.}
}
\description{
Create a labelled vector with defined unit of measure
}
