% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mall.R
\docType{data}
\name{mall}
\alias{mall}
\title{Mall_Customers Dataset}
\format{A data frame with 200 rows and 5 variables:
\describe{
\item{\code{CustomerID}}{double. DESCRIPTION.}
\item{\code{Gender}}{character. DESCRIPTION.}
\item{\code{Age}}{double. DESCRIPTION.}
\item{\code{Annual Income (k$)}}{double. DESCRIPTION.}
\item{\code{Spending Score (1-100)}}{double. DESCRIPTION.}
}}
\usage{
mall
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
