% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_contributor.R
\name{map_role_to_schema}
\alias{map_role_to_schema}
\title{Map R person roles to schema.org-style roles}
\usage{
map_role_to_schema(role)
}
\arguments{
\item{role}{A character vector of roles (e.g. \code{"cre"}, \code{"ctb"}).}
}
\value{
A character vector with schema.org-style roles.
}
\description{
Map R person roles to schema.org-style roles
}
\keyword{internal}
