\name{bivan-methods}
\docType{methods}
\alias{bivan-methods}
\alias{bivan,formula,Dataset-method}
\title{ Bivariate analysis measures computation }
\description{
\code{bivan} computes some main bivariate analysis measures.
}
\section{Methods}{
\describe{

\item{\code{signature(formula = "formula", data = "Dataset")}}{
\code{bivan} method for a \code{Dataset} object.
}
}}
\keyword{methods}
\keyword{bivan}
\keyword{bivariate}
\keyword{analysis}
