\name{summaryToPDF-methods}
\docType{methods}
\alias{summaryToPDF-methods}
\alias{summaryToPDF,Bivan-method}
\alias{summaryToPDF,Dataset-method}
\alias{summaryToPDF,Statdf-method}
\title{Produce a summary in a PDF file}
\description{
\code{summaryToPDF} produces a summary of the object in a PDF file.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Bivan")}}{
\code{summaryToPDF} method for a \code{Bivan} object.
}

\item{\code{signature(object = "Dataset")}}{
\code{summaryToPDF} method for a \code{Dataset} object.
}

\item{\code{signature(object = "Statdf")}}{
\code{summaryToPDF} method for a \code{Statdf} object.
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
