\name{varid-methods}
\docType{methods}
\alias{varid-methods}
\alias{varid,character,Dataset-method}
\title{Get the column id of given variable names}
\description{
\code{varid} returns the column ids of one or more variable names of an object storing variables.
}
\section{Methods}{
\describe{

\item{\code{signature(names = "character", object = "Dataset")}}{
\code{varid} method for a \code{Dataset} object. 
}
}}
\keyword{methods}
\keyword{index}
\keyword{column}
