% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining_samples.R
\docType{data}
\name{mining_samples}
\alias{mining_samples}
\title{Mining_Process Dataset sample}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 24 columns.}
\usage{
mining_samples
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
