% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_triple.R
\name{n_triples}
\alias{n_triples}
\title{Create N-Triples}
\usage{
n_triples(triples)
}
\arguments{
\item{triples}{A character vector of concatenated N-Triples, created with
\code{\link[=n_triple]{n_triple()}}.}
}
\value{
A character vector of unique N-Triple strings.
}
\description{
Create RDF triple statements to annotate your dataset with standard,
interoperable metadata.
}
\details{
N-Triples is a line-based serialization format for RDF. It is easy to parse
and widely supported. For details, see the
\href{https://www.w3.org/TR/rdf12-n-triples/}{W3C RDF 1.2 N-Triples specification}.
}
\examples{
triple_1 <- n_triple(
  "http://example.org/show/218",
  "http://www.w3.org/2000/01/rdf-schema#label",
  "That Seventies Show"
)

triple_2 <- n_triple(
  "http://example.org/show/218",
  "http://example.org/show/localName",
  '"Cette S<U+00E9>rie des Ann<U+00E9>es Septante"@fr-be'
)

n_triples(c(triple_1, triple_2, triple_1))

}
