\name{QuantitativeVariable-class}
\Rdversion{1.1}
\docType{class}
\alias{QuantitativeVariable-class}

\title{Class \code{"QuantitativeVariable"}}
\description{
An object for storing quantitative variables.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{codes}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{values}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{missings}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{description}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{Variable}"}, directly.
}
\section{Methods}{
No methods defined with class "QuantitativeVariable" in the signature.
}

\author{
Emmanuel Rousseaux
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("QuantitativeVariable")
}
\keyword{classes}
