% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_function.R
\name{read_dataset}
\alias{read_dataset}
\title{Read a file into a dataset}
\usage{
read_dataset(
  dataset_id,
  obs_id = NULL,
  Title,
  dimensions,
  measurements,
  attributes,
  unit,
  .f = "utils::read.csv",
  ...
)
}
\arguments{
\item{obs_id}{The observation identifier.}

\item{Title}{The title of the dataset, corresponds to datacite:Title, dct:title, rdfs:label.}

\item{dimensions}{The columns in the data frame that are dimensions, and maybe used
for statistical aggregation.}

\item{measurements}{The measurements in the data frame.}

\item{attributes}{Not measured attributes of the observations}

\item{unit}{The unit of the measurements}

\item{.f}{A function (from an installed package)}

\item{...}{Parameters to read the file.}
}
\value{
A dataset with a recorded Time and RelatedItem attribute.
}
\description{
Read a file into a dataset
}
\examples{
temp_path <- file.path(tempdir(), "iris.csv")
write.csv(iris, file = temp_path, row.names = F)
iris_ds <- read_dataset(
  dataset_id = "iris_dataset",
  obs_id = NULL,
  dimensions = NULL,
  measurements = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),
  attributes = "Species",
  Title = "Iris Dataset",
  unit = list(code = "MM", label = "millimeter"),
  .f = "utils::read.csv", temp_path )
attributes(iris_ds)
}
