% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_item.R
\name{related_item}
\alias{related_item}
\title{Create a related item}
\usage{
related_item(
  related_item_identifier,
  creator,
  title,
  publication_year,
  volume = NULL,
  issue = NULL,
  number = NULL,
  number_type = NULL,
  first_page = NULL,
  last_page = NULL,
  publisher = NULL,
  edition = NULL,
  contributor = NULL
)
}
\arguments{
\item{related_item_identifier}{A list created with \code{\link{related_item_identifier}}.}
}
\value{
a related item.
}
\description{
Create a related item
}
\seealso{
Other metadata functions: 
\code{\link{datacite}()},
\code{\link{dublincore}()},
\code{\link{related_item_identifier}()},
\code{\link{size_add}()}
}
\concept{metadata functions}
