% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retail.R
\docType{data}
\name{retail}
\alias{retail}
\title{Online_Retail Dataset}
\format{A data frame with 541909 rows and 8 variables:
\describe{
\item{\code{InvoiceNo}}{character. DESCRIPTION.}
\item{\code{StockCode}}{character. DESCRIPTION.}
\item{\code{Description}}{character. DESCRIPTION.}
\item{\code{Quantity}}{double. DESCRIPTION.}
\item{\code{InvoiceDate}}{double. DESCRIPTION.}
\item{\code{UnitPrice}}{double. DESCRIPTION.}
\item{\code{CustomerID}}{double. DESCRIPTION.}
\item{\code{Country}}{character. DESCRIPTION.}
}}
\usage{
retail
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
