% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-statwales.R
\docType{data}
\name{statwales}
\alias{statwales}
\title{Life Expectancy in Regions of Wales by Sex}
\format{
The dataset has originally one column for reference areas, and three
column-pairs in slices, in total 7 column, and 4 reference area rows.
}
\source{
\href{https://www.w3.org/TR/vocab-data-cube/#dataset}{The RDF Data Cube Vocabulary},
example in 5.4.
}
\usage{
statwales
}
\description{
A small demonstration data set extracted from StatsWales report
number 003311 which describes life expectancy broken down by region
(unitary authority), age and time.
}
\keyword{datasets}
