% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjecst.R
\name{subjects}
\alias{subjects}
\alias{subject_add}
\title{Add subject(s) to a dataset}
\usage{
subjects(x)

subject_add(x, term, scheme, identifier)
}
\arguments{
\item{x}{An R object}

\item{term}{A term, or a character vector of multiple terms.}

\item{scheme}{The scheme to which the term correspondes. If there are multiple terms,
provide the scheme(s) in the same order.}

\item{identifer}{The identifier of the term definition(s), in the order of the terms.}
}
\value{
The subjects as a data.frame of terms
}
\description{
Add one or more subject terms to the dataset's metadata.
}
\details{
In the Dublin Core elements, \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#http://purl.org/dc/elements/1.1/subject}{dct::subject} is
defined Typically, the subject will be represented using keywords, key phrases, or classification codes. It is
recommended as a best practice to use a controlled vocabulary. \cr
In DataCite, subjects are defined as key phrases from a controlled library.
}
\examples{
x <- data.frame()
subject_add (x, term = c("R (Computer program language)",
                         "Questionnaires--Computer programs"),
                scheme = rep("url", 2),
                identifier = c("https://id.loc.gov/authorities/subjects/sh2002004407.html",
                               "http://id.worldcat.org/fast/1085693/") )
 subjects(x)
}
