% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_function.R
\name{use_function}
\alias{use_function}
\title{Use a function on a dataset}
\usage{
use_function(dataset, .f = "base::subset", ...)
}
\arguments{
\item{dataset}{A dataset object.}

\item{.f}{A function (from an installed package)}
}
\value{
A dataset with a recorded Time and RelatedItem attribute.
}
\description{
Use a function on a dataset
}
\examples{
my_dataset <- dataset (x = data.frame ( geo = c("NL", "NL", "BE", "BE"),
                       values = c(1:4)),
                       dataset_id = "reprex",
                       obs_id = NULL,
                       dimensions = "geo",
                       measurements = "values",
                       attributes = NULL,
                       Title =  "Sample Dataset")

 attributes(use_function(my_dataset, .f="summary"))
}
