% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{version_add}
\alias{version_add}
\title{Add the version of the object.}
\usage{
version_add(x, version)
}
\arguments{
\item{x}{An R object, such as a data.frame, a tibble, or a character vector.}

\item{version}{The version as a character set.}
}
\value{
The Version attribute as a character of length 1 is added to \code{x}.
}
\description{
Add the optional Version property as an attribute to an R object.
}
\details{
Version is an optional property in
\href{https://support.datacite.org/docs/schema-optional-properties-v43#13-size}{DataCite 4.3}.
It is not part of the "core" Dublin Core terms, but ...
\href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/}{Dublin Core metadata terms}.
}
\examples{
iris_dataset <- version_add(x = iris, version= "1.0")
attr(iris_dataset, "Version")
}
\seealso{
Other Optional metadata: 
\code{\link{size_add}()}
}
\concept{Optional metadata}
