% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defined.R
\name{as_character}
\alias{as_character}
\alias{as_character.haven_labelled_defined}
\title{Coerce to character vector}
\usage{
as_character(x)

\method{as_character}{haven_labelled_defined}(x)
}
\arguments{
\item{x}{A vector created with \code{\link{defined}}.}
}
\value{
A character vector.
}
\description{
Coerce to character vector
}
\examples{
as_character(iris_dataset$Species)
}
