% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defined.R
\name{as_factor}
\alias{as_factor}
\title{Coerce to factor vector}
\usage{
as_factor(x)
}
\arguments{
\item{x}{A vector created with \code{\link{defined}}.}

\item{...}{Further arguments passed to internal methods (not used).}
}
\value{
A factor vector.
}
\description{
Coerce to factor vector
}
\examples{
sex <- defined(
  c(0, 1, 1, 0),
  label = "Sex",
  labels = c("Female" = 0, "Male" = 1)
)
as_factor(sex)
}
