% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defined.R
\name{as_numeric}
\alias{as_numeric}
\alias{as_numeric.haven_labelled_defined}
\title{Coerce a defined vector to numeric}
\usage{
as_numeric(x)

\method{as_numeric}{haven_labelled_defined}(x)
}
\arguments{
\item{x}{A vector created with \code{\link{defined}}.}
}
\value{
A numeric vector.
}
\description{
Coerce a defined vector to numeric
}
\examples{
as_numeric(iris_dataset$Petal.Length)
}
