% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as_tibble.dataset_df}
\alias{as_tibble.dataset_df}
\alias{as.tibble.dataset_df}
\title{Convert a dataset_df to a tibble}
\usage{
\method{as_tibble}{dataset_df}(x, ..., strip_attributes = TRUE)

as.tibble.dataset_df(x, ..., strip_attributes = TRUE)
}
\arguments{
\item{x}{A \code{\link{dataset_df}}.}

\item{...}{Passed to \code{base::as.data.frame()}.}

\item{strip_attributes}{Logical: should column-level semantic metadata
be stripped? Default: \code{TRUE}.}
}
\value{
A tibble with optional metadata stripping.
}
\description{
A thin wrapper around \code{as.data.frame.dataset_df()},
returning a tibble.
}
