% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_defined_rows.R
\name{bind_defined_rows}
\alias{bind_defined_rows}
\title{Bind strictly defined rows}
\usage{
bind_defined_rows(x, y, ...)
}
\arguments{
\item{x}{A dataset created with \code{\link{dataset_df}}.}

\item{y}{A dataset created with \code{\link{dataset_df}}.}

\item{...}{Optional parameters: \code{dataset_title}, \code{creator}.}
}
\value{
A dataset with the combined rows.
}
\description{
Add rows of the y dataset to the x dataset.
}
\details{
By default, the dataset_bibentry bibliographical data and the
title is recycled from dataset x. You can give a new title with
..., title="New Title". \cr
By default, the unique creators of dataset y, who are not present in
dataset x, are added to the creators of the new dataset.
}
\examples{
A <- dataset_df(
  a = defined(c(11, 14, 16), label = "length", unit = "cm"),
  dataset_bibentry = dublincore(
    title = "Test", creator = person("Jane Doe"),
    dataset_date = Sys.Date()
  )
)

B <- dataset_df(
  a = defined(c(12, 17, 19), label = "length", unit = "cm"),
  dataset_bibentry = dublincore(
    title = "Test", creator = person("Jane Doe")
  )
)
bind_defined_rows(x = A, y = B)

}
