% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator.R
\name{creator}
\alias{creator}
\alias{creator<-}
\title{Get/set the Creator of the object.}
\usage{
creator(x)

creator(x, overwrite = TRUE) <- value
}
\arguments{
\item{x}{A semantically rich data frame object created by
\code{\link[=dataset_df]{dataset_df()}} or \verb{dataset::\link{as_dataset_df}}.}

\item{overwrite}{If the attributes should be overwritten. In case it is set
to \code{FALSE},it gives a message with the current \code{Creator} property
instead of overwriting it. Defaults to \code{TRUE} when the attribute is
set to \code{value} regardless of previous setting.}

\item{value}{The \code{Creator} as a
\code{\link[utils:person]{utils::person()}} object.}
}
\value{
The Creator attribute as a character of length one is added to
\code{x}.
}
\description{
Add the optional \code{Creator} property as an attribute to a
dataset object.
}
\details{
The \code{Creator} corresponds to
\href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/creator/}{dct:creator}
in Dublin Core and Creator in DataCite. The name of the entity that holds,
archives, publishes prints, distributes, releases, issues, or produces the
dataset. This property will be used to formulate the citation, so consider
the prominence of the role.
}
\examples{
creator(orange_df)
# To change author:
creator(orange_df) <- person("Jane", "Doe")
# To add author:
creator(orange_df, overwrite = FALSE) <- person("John", "Doe")
}
\seealso{
Other bibliographic helper functions:
\code{\link[=contributor]{contributor()}},
\code{\link[=dataset_format]{dataset_format()}},
\code{\link[=dataset_title]{dataset_title()}},
\code{\link[=description]{description()}},
\code{\link[=geolocation]{geolocation()}},
\code{\link[=get_bibentry]{get_bibentry()}},
\code{\link[=language]{language()}},
\code{\link[=publication_year]{publication_year()}},
\code{\link[=publisher]{publisher()}},
\code{\link[=relation]{relation()}},
\code{\link[=rights]{rights()}},
\code{\link[=subject]{subject()}}
}
\concept{bibliographic helper functions}
