\name{dds}
\alias{dds}
\docType{data}
\title{
Data sample for the Dataset package
}
\description{
A small data base for testing purpose only. These data are under license, you can't use it in your work. Thank you for your comprehension.
}
\usage{data(dds)}
\format{
A Dataset object with 15 rows and 10 columns.
}
\details{
Please run a summaryToPDF.
}
\source{
NA
}
\references{
NA
}
\examples{
data(dds)
summaryToPDF(dds, description.chlength = 60)
}
\keyword{datasets}
\keyword{data}
\keyword{Dataset}
\keyword{dds}
