% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{datacite}
\alias{datacite}
\title{Add DataCite metadata to a data frame}
\source{
\href{https://support.datacite.org/docs/schema-mandatory-properties-v43}{DataCite 4.3 Mandatory Properties} and
\href{https://support.datacite.org/docs/schema-optional-properties-v43}{DataCite 4.3 Optional Properties}
}
\usage{
datacite(
  df,
  Title,
  Creator,
  Publisher,
  PublicationYear = "THIS",
  Subject = NULL,
  Contributor = NULL,
  Date = NULL,
  Language = NULL,
  AlternateIdentifer = NULL,
  RelatedIdentifier = NULL,
  Format = NULL,
  Version = NULL,
  Rights = NULL,
  Description = NULL,
  Geolocation = NULL,
  FundingReference = NULL
)
}
\arguments{
\item{df}{An R object of type data.frame, or inherited data.table, tibble; alternatively a well
structured R list.}

\item{Creator}{The main researchers involved in producing the data, or the authors of the publication, in priority order. To supply multiple creators, repeat this property.}

\item{PublicationYear}{The year when the data was or will be made publicly available in \code{YYYY} format.}

\item{Subject}{Recommended for discovery. Subject, keyword, classification code, or key phrase describing the resource.}

\item{Contributor}{Recommended for discover. The institution or person responsible for collecting, managing, distributing, or otherwise contributing to the development of the resource.}

\item{Date}{Recommended for discovery.}

\item{Language}{The primary language of the resource. Allowed values are taken from IETF BCP 47, ISO 639-1 language.}

\item{RelatedIdentifier}{Recommended for discovery.}

\item{Format}{Technical format of the resource.}

\item{Version}{Free text. Suggested practice: track major_version.minor_version.}

\item{Rights}{Any rights information for this resource. The property may be repeated to record complex rights characteristics.
Free text.}

\item{Description}{Recommended for discovery. All additional information that does not fit in any of the other categories. May be used for technical information. A free text.}

\item{Geolocation}{Recommended for discovery. Spatial region or named place where the data was gathered or about which the data is focused.}

\item{FundingReference}{Information about financial support (funding) for the resource being registered.}

\item{Pulisher}{The name of the entity that holds, archives, publishes prints, distributes, releases, issues, or produces the resource. This property will be used to formulate the citation, so consider the prominence of the role. For software, use Publisher for the code repository. If there is an entity other than a code repository, that "holds, archives, publishes, prints, distributes, releases, issues, or produces" the code, use the property Contributor/contributorType/ hostingInstitution for the code repository.}

\item{AlternateIdentifier}{An identifier or identifiers other than the primary Identifier applied to the resource being registered. This may be any alphanumeric string which is unique within its domain of issue. May be used for local identifiers. AlternateIdentifier should be used for another identifier of the same instance (same location, same file).}
}
\value{
An R object with at least the mandatory DataCite attributes.
}
\description{
Add metadata conforming the \href{https://schema.datacite.org/}{DataCite Metadata Schema}
to datasets, i.e. structured R data.frame or list objects, for an accurate and consistent identification
of a resource for citation and retrieval purposes.
}
\details{
DataCite is a leading global non-profit organisation that provides persistent identifiers
(DOIs) for research data and other research outputs. Organizations within the research
community join DataCite as members to be able to assign DOIs to all their research
outputs. This way, their outputs become discoverable and associated metadata is made
available to the community.
DataCite then develops additional services to improve the DOI management experience,
making it easier for our members to connect and share their DOIs with the broader
research ecosystem and to assess the use of their DOIs within that ecosystem.
DataCite is an active participant in the research community and promotes data sharing
and citation through community-building efforts and outreach activities.

The \code{ResourceType} property will be by definition "Dataset".
The \code{Size} attribute (e.g. bytes, pages, inches, etc.) will automatically added to the dataset.
}
\examples{
datacite_dataset(df = iris,
                 Title = "Iris Dataset",
                 Creator = person("Anderson", "Edgar", role = "aut"),
                 Publisher=" American Iris Society",
                 PublicationYear = 1935,
                 geoLocation = "US",
                 Language = "en")
}
\seealso{
Other metadata functions: 
\code{\link{dataset}()}
}
\concept{metadata functions}
