% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_df.R
\name{dataset_df}
\alias{dataset_df}
\alias{print.dataset_df}
\alias{is_dataset_df}
\alias{names.dataset_df}
\title{Create a new dataset_df object}
\usage{
dataset_df(reference = NULL, var_labels = NULL, units = NULL, ...)

\method{print}{dataset_df}(x, ...)

is_dataset_df(x)

\method{names}{dataset_df}(x)
}
\arguments{
\item{reference}{A list of bibliographic references and descriptive metadata
about the dataset as a whole.}

\item{var_labels}{The long, human readable labels of each variable.}

\item{units}{The units of measurement for the measured variables.}

\item{...}{The vectors (variables) that should be included in the dataset.}
}
\value{
A dataset_df object with rich metadata.
}
\description{
Create a new dataset_df object
}
