% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_df.R
\name{dataset_df}
\alias{dataset_df}
\alias{as_dataset_df}
\alias{is.dataset_df}
\alias{print.dataset_df}
\alias{is_dataset_df}
\title{Create a new dataset_df object}
\usage{
dataset_df(
  reference = NULL,
  var_labels = NULL,
  units = NULL,
  definitions = NULL,
  ...
)

as_dataset_df(
  df,
  var_labels = NULL,
  units = NULL,
  definitions = NULL,
  reference = NULL,
  ...
)

is.dataset_df(x)

\method{print}{dataset_df}(x, ...)

is_dataset_df(x)
}
\arguments{
\item{reference}{A list of bibliographic references and descriptive metadata
about the dataset as a whole.}

\item{var_labels}{The long, human readable labels of each variable.}

\item{units}{The units of measurement for the measured variables.}

\item{definitions}{The linked definitions of the variables, attributes, or constants.}

\item{...}{The vectors (variables) that should be included in the dataset.}

\item{df}{A data.frame to be converted to dataset_df.}

\item{x}{A dataset for S3 methods.}
}
\value{
A dataset_df object with rich metadata.
}
\description{
Create a new dataset_df object
}
\details{
To check if an object is a dataset_df use \code{is.dataset_df(x)}.
}
