% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_title.R
\name{dataset_title}
\alias{dataset_title}
\alias{title_add}
\title{Read / Add DataCite metadata to an object}
\usage{
dataset_title(x)

title_add(x, Title, titleType = NULL)
}
\arguments{
\item{x}{An R object of type data.frame, or inherited data.table, tibble; alternatively a well
structured R list.}

\item{Title}{The name(s) or title(s) by which a resource is known. May be the title of a dataset
or the name of a piece of software. Similar to \href{http://purl.org/dc/elements/1.1/title}{dct:title}.\cr
See \code{\link{dataset_title}} for adding further titles.}

\item{titleType}{For a single \code{Title} defaults to \code{NULL}. Otherwise you can add
a Subtitle, an Alternative Title and an Other Title. See \code{\link{dataset_title}}.}
}
\description{
Read/Add a Title attribute to a dataset or R object
}
\details{
For Dublin Core metadata, use Title as a character vector of length 1, with
titleType = \code{NULL}.\cr
For DataCite, provide a vector of titles and an equal length of titleType parameters.
}
\examples{
my_iris <- iris
my_iris <- title_add(my_iris,
                     Title = c("The Famous Dataset", "Iris Dataset"),
                      titleType = c("AlternativeTitle", "Title")
                     )
dataset_title(my_iris)
}
