% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_title.R
\name{dataset_title}
\alias{dataset_title}
\alias{dataset_title<-}
\title{Get/set title of a dataset}
\usage{
dataset_title(x)

dataset_title(x, overwrite = FALSE) <- value
}
\arguments{
\item{x}{A dataset object created with \code{dataset::\link{dataset}}.}

\item{value}{The name(s) or title(s) by which a resource is known. See:
\href{https://purl.org/dc/elements/1.1/title}{dct:title}.}
}
\value{
A string with the dataset's title; \code{set_dataset_title} returns
a dataset object with the changed (main) title.
}
\description{
Get or reset the dataset's main title.
}
\details{
In the DataCite definition, several titles can be used; it is not
yet implemented.
}
\examples{
dataset_title(iris_dataset)
dataset_title(iris_dataset, overwrite = TRUE) <-"The Famous Iris Dataset"
dataset_title(iris_dataset)
}
