% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_to_triples.R
\name{dataset_to_triples}
\alias{dataset_to_triples}
\title{Dataset to triples (three columns)}
\usage{
dataset_to_triples(df, idcol = NULL)
}
\arguments{
\item{df}{A \link{data.frame} or similar object, or a \link{dataset}.}

\item{idcol}{The identifier column. If \code{idcol} is \code{NULL} it attempts to
use the \code{row.names(df)} as an \code{idcol}.}
}
\value{
The long form version of the original dataset, retaining the attributes
and class.
}
\description{
The dataset is converted into a three-column long format with
columns \code{s} for subject, \code{p} for predicate and \code{o} for
object.
}
