% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_to_triples.R
\name{dataset_to_triples}
\alias{dataset_to_triples}
\title{Dataset to triples (three columns)}
\usage{
dataset_to_triples(x, idcol = NULL)
}
\arguments{
\item{x}{An R object that contains the data of the dataset (a data.frame or
inherited from \code{\link[base:data.frame]{data.frame}}), for example, \code{\link[=dataset_df]{dataset_df()}}
\code{\link[tibble:tibble]{tibble::tibble()}}, \code{\link[tsibble:tsibble]{tsibble::tsibble()}}, \code{\link[data.table:data.table]{data.table::data.table()}}.}

\item{idcol}{The identifier column. If \code{idcol} is \code{NULL} it attempts to
use the \code{row.names(df)} as an \code{idcol}.}
}
\value{
The long form version of the original dataset, retaining the attributes
and class.
}
\description{
The dataset is converted into a three-column long format with
columns \code{s} for subject, \code{p} for predicate and \code{o} for
object.
}
