% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defined.R
\name{defined}
\alias{defined}
\alias{is.defined}
\alias{as.character.haven_labelled_defined}
\alias{summary.haven_labelled_defined}
\title{Create a semantically well-defined, labelled vector}
\usage{
defined(
  x,
  labels = NULL,
  label = NULL,
  unit = NULL,
  definition = NULL,
  namespace = NULL
)

is.defined(x)

\method{as.character}{haven_labelled_defined}(x, ...)

\method{summary}{haven_labelled_defined}(object, ...)
}
\arguments{
\item{x}{A vector to label. Must be either numeric (integer or double) or
character.}

\item{labels}{A named vector or \code{NULL}. The vector should be the same type
as \code{x}. Unlike factors, labels don't need to be exhaustive: only a fraction
of the values might be labelled.}

\item{label}{A short, human-readable description of the vector or \code{NULL}.}

\item{unit}{A character string of length one containing the
unit of measure or \code{NULL}.}

\item{definition}{A character string of length one containing a
linked definition or \code{NULL}.}

\item{namespace}{A namespace for individual observations or categories or \code{NULL}.}

\item{...}{Further parameters for inheritance, not in use.}

\item{object}{An R object to be summarised.}
}
\value{
The constructor \code{defined} returns a vector with defined
value labels, a variable label, an optional unit of measurement and linked
definition.
}
\description{
Create a semantically well-defined, labelled vector
}
