% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describe.dataset}
\title{Describe a dataset object}
\usage{
describe(x)

\method{describe}{dataset}(x)
}
\arguments{
\item{x}{A dataset object created with \code{dataset::\link{dataset}}.}
}
\value{
No object is returned, but they key attributes are printed on the screen.
}
\description{
A convenience function to review the most important attributes of a dataset
object. The dataset class adds a wide range of metadata as attributes to
data.frame, tibble, data.table or similar R object that contains tabular data.
Overviewing these attributes becomes cumbersome with base
R \code{attributes()}.
}
\examples{
describe(iris_dataset)

}
