% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe a dataset}
\usage{
describe(x, con = NULL)
}
\arguments{
\item{x}{A dataset_df object.}

\item{con}{A connection, for example, \code{con=tempfile()}.}
}
\value{
The description of the dataset_df object is written to the
connection in the N-Triples form. If \code{con=NULL}, then the serialisation
takes place in \code{tempfile()} and the contents are printed to the console;
if a file is given, than no output is returned.
}
\description{
Describe a dataset
}
\examples{

# See the serialisation on the screen:
describe(orange_df)

# Save it to a connection:
temporary_connection <- tempfile()
describe(orange_df, con = temporary_connection)
}
