% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe a dataset in N-Triples format}
\usage{
describe(x, con)
}
\arguments{
\item{x}{A \code{dataset_df} object.}

\item{con}{A connection or a character string path (e.g. from \code{tempfile()}).}
}
\value{
Invisibly returns \code{x}. Writes N-Triples to \code{con}.
}
\description{
Writes provenance and Dublin Core metadata of a dataset to a
file or connection in N-Triples format.
}
