% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe a dataset}
\usage{
describe(x, con)
}
\arguments{
\item{x}{A dataset_df object.}

\item{con}{A connection, for example, \code{con=tempfile()}.}
}
\value{
The description of the dataset_df object is written to the connection
in the n-triples form, nothing is returned.
}
\description{
Describe a dataset
}
\examples{
temp_prov <- tempfile()
describe(iris_dataset, con=temp_prov)
readLines(temp_prov)
}
