% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\alias{description<-}
\title{Get/set the Description of the object.}
\usage{
description(x)

description(x, overwrite = FALSE) <- value
}
\arguments{
\item{x}{A dataset object created with \code{dataset::\link{dataset_df}} or
\code{dataset::\link{as_dataset_df}}.}

\item{overwrite}{If the \code{Description} attribute should be overwritten.
In case it is set to \code{FALSE}, it gives a message with the current
\code{Description} property instead of overwriting it. Defaults to
\code{FALSE}, when it gives a warning at an accidental overwrite attempt.}

\item{value}{The \code{Description} as a character set.}
}
\value{
The \code{Description} attribute as a character of length 1 is added
to \code{x}.
}
\description{
Get/set the optional \code{Description} property as an attribute
to an R object.
}
\details{
The \code{Description} is recommended for discovery in DataCite. All
additional information that does not fit in any of the other categories.
May be used for technical information. A free text. Similar to
\href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/description/}{dct:description}.
}
\examples{
description(orange_df)
description(
  orange_df,
  overwrite = TRUE
) <- "The 'orange' dataset has 35 rows and 3 columns
                          of records of the growth of orange trees."
description(orange_df)
}
\seealso{
Other Reference metadata functions: 
\code{\link{geolocation}()},
\code{\link{identifier}()},
\code{\link{language}},
\code{\link{publication_year}()},
\code{\link{publisher}()},
\code{\link{rights}()}
}
\concept{Reference metadata functions}
