% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bibentry.R
\name{get_bibentry}
\alias{get_bibentry}
\alias{set_bibentry}
\title{Get the bibentry item of a dataset}
\usage{
get_bibentry(dataset)

set_bibentry(dataset, value)
}
\arguments{
\item{dataset}{A dataset created with \code{\link{dataset_df}}.}
}
\value{
A \code{\link[utils]{bibentry}} bibliographic entry created for the dataset.
}
\description{
Get the bibentry item of a dataset
}
\examples{
get_bibentry(iris_dataset)
}
