% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bibentry.R
\name{get_bibentry}
\alias{get_bibentry}
\alias{set_bibentry}
\title{Get/set the Bibentry of the object.}
\usage{
get_bibentry(dataset)

set_bibentry(dataset, value)
}
\arguments{
\item{dataset}{A dataset created with \code{\link{dataset_df}}.}

\item{value}{A \code{\link[utils:bibentry]{utils::bibentry}} object, or a newly
initialised bibentry object with DataCite default values for unassigned entries.}
}
\value{
A \code{\link[utils]{bibentry}} bibliographic entry created for the dataset.
}
\description{
Get/set the Bibentry of the object.
}
\examples{
get_bibentry(iris_dataset)
}
