% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_column.R
\name{id_to_column}
\alias{id_to_column}
\title{Add identifier to columns}
\usage{
id_to_column(x, prefix = "eg:", ids = NULL)
}
\arguments{
\item{x}{A dataset created with \code{\link{dataset_df}}.}

\item{prefix}{Defaults to \code{eg:} (example.com).}

\item{ids}{Defaults to \code{NULL}.}
}
\value{
A dataset conforming the original sub-class of \code{x}.
}
\description{
Add a prefixed identifier to the first column of the dataset.
}
\examples{

# Example with a dataset_df object:
id_to_column(iris_dataset)

# Example with a data.frame object:#'
id_to_column(iris, prefix="eg:iris-o")
}
