% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{measures}
\alias{measures}
\alias{measures_add}
\title{Measures of a dataset}
\usage{
measures(x)

measures_add(x, measures)
}
\arguments{
\item{x}{A data.frame or inherited tibble, data.frame, or a structured list.}

\item{measures}{The name or column number of the measures within the dataset.}

\item{sdmx_attributes}{The optional SDMX dimensions.}
}
\description{
Measures of a dataset
}
\details{
See the W3C and SDMX definition of a \href{https://www.w3.org/TR/vocab-data-cube/#dsd-dimensions}{measure}.
}
\examples{
df <- data.frame ( sex = c("M", "F"), value = c(1,2))
df <- measures_add(df, "value")
measures(df)
}
