\name{bivan-methods}
\docType{methods}
\alias{bivan-methods}
\alias{bivan,formula,Dataset-method}
\title{Compute some bivariate analysis}
\description{
\code{bivan} returns the results of the bivariate analysis requested by the user.
}
\section{Methods}{
\describe{

\item{\code{signature(formula = "formula", data = "Dataset")}}{
\code{bivan} method for a \code{Dataset} object.
}
}}
\keyword{methods}
\keyword{bivariate}
