% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{provenance}
\alias{provenance}
\title{Provenance}
\usage{
provenance(dataset)
}
\arguments{
\item{dataset}{A dataset created with \code{\link{dataset_df}}.}
}
\value{
A string with N-Triple format provenance metadata
}
\description{
Provenance
}
\examples{
provenance(iris_dataset)
}
