% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{provenance}
\alias{provenance}
\alias{provenance<-}
\title{Get or update provenance information}
\usage{
provenance(x)

provenance(x) <- value
}
\arguments{
\item{value}{A list that may contain the following elements: \code{wasInformedBy},
\code{wasAssociatedWith}.}
}
\description{
Add or update information about the history (provenance) of the dataset.
}
\details{
For additional details see \code{vignette("provenance", package = "dataset")}.
}
\examples{
# Get the provenance of a dataset:
provenance(iris_dataset)

# Update the provenance:
provenance(iris_dataset) <- list(
          wasInformedBy="https://doi.org/10.1111/j.1469-1809.1936.tb02137.x"
          )
}
