% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{provenance}
\alias{provenance}
\alias{provenance<-}
\title{Get or update provenance information}
\usage{
provenance(x)

provenance(x) <- value
}
\arguments{
\item{x}{An R object that contains the data of the dataset (a data.frame or
inherited from \code{\link[base:data.frame]{data.frame}}), for example,
\code{\link[tibble:tibble]{tibble::tibble()}}, \code{\link[tsibble:tsibble]{tsibble::tsibble()}}, \code{\link[data.table:data.table]{data.table::data.table()}}.}

\item{value}{A list that may contain the following elements: \code{wasInformedBy},
\code{wasAssociatedWith}.}
}
\description{
Get or update provenance information
}
\examples{
# Get the provenance of a dataset:
provenance(iris_dataset)

# Update the provenance:
provenance(iris_dataset) <- list(
          wasInformedBy="https://doi.org/10.1111/j.1469-1809.1936.tb02137.x"
          )
}
