% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_item.R
\name{related_item_identifier}
\alias{related_item_identifier}
\title{Create a related item identifer}
\usage{
related_item_identifier(scheme, identifier, relation_type, resource_type)
}
\arguments{
\item{scheme}{See, recommended to use \code{"URL"} or \code{"DOI"}.}

\item{identifier}{The unique identifier of the resource.}

\item{relation_type}{Seee}

\item{resource_type}{Seee}
}
\value{
A list
}
\description{
Create a related item identifer
}
\examples{
related_item_identifer
}
\seealso{
Other metadata functions: 
\code{\link{datacite}()},
\code{\link{dublincore}()},
\code{\link{related_item}()},
\code{\link{size_add}()}
}
\concept{metadata functions}
