% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{subject}
\alias{subject}
\alias{subject_create}
\alias{subject<-}
\alias{is.subject}
\title{Create, add, or retrieve a subject}
\usage{
subject(x)

subject_create(
  term,
  schemeURI = NULL,
  valueURI = NULL,
  prefix = NULL,
  subjectScheme = NULL,
  classificationCode = NULL
)

subject(x) <- value

is.subject(x)
}
\arguments{
\item{x}{A dataset object created with \code{\link[=dataset_df]{dataset_df()}} or \code{\link[=as_dataset_df]{as_dataset_df()}}.}
}
\value{
\itemize{
\item \code{subject(x)} returns the \code{"subject"} attribute or the \code{subject} field
from the dataset's bibentry.
\item \code{subject(x) <- value} sets the \code{"subject"} attribute and updates the
bibentry's \code{subject} field, returning the modified dataset invisibly.
}
}
\description{
Manage the subject metadata of a dataset. The subject can be a simple
character term or a structured object with subproperties created by
\code{\link[=subject_create]{subject_create()}}.
}
\details{
The \href{https://schema.datacite.org/meta/kernel-4/}{DataCite subject property}
supports multiple subproperties. These cannot be stored directly in a
standard \code{\link[utils:bibentry]{utils::bibentry}} object.
Therefore:
\itemize{
\item If you set a character string as the subject, it is stored in the bibentry
and as a \code{"subject"} attribute.
\item If you set a structured subject (via \code{\link[=subject_create]{subject_create()}}), the \verb{$term}
value is stored in the bibentry, and the complete object is stored in the
\code{"subject"} attribute of the \code{\link{dataset_df}} object.
}
}
\examples{
df <- dataset_df(data.frame(x = 1))

# Set subject as a structured object
subject(df) <- subject_create(
  term = "Oranges",
  schemeURI = "http://id.loc.gov/authorities/subjects",
  valueURI = "http://id.loc.gov/authorities/subjects/sh85095257",
  subjectScheme = "LCCH",
  prefix = "lcch:"
)

# Retrieve subject
subject(df)

}
\seealso{
Other bibliographic helper functions: 
\code{\link{contributor}()},
\code{\link{creator}()},
\code{\link{dataset_title}()},
\code{\link{description}()},
\code{\link{geolocation}()},
\code{\link{get_bibentry}()},
\code{\link{language}},
\code{\link{publication_year}()},
\code{\link{publisher}()},
\code{\link{rights}()}
}
\concept{bibliographic helper functions}
