% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject.R
\name{subject}
\alias{subject}
\alias{subject<-}
\alias{is.subject}
\alias{is.subject.subject}
\title{Create a subject}
\usage{
subject(ds)

subject(ds)

subject(ds) <- value

is.subject(x)

\method{is.subject}{subject}(x)
}
\arguments{
\item{ds}{A dataset which has or needs a subject field in the metadata
attributes.}

\item{x}{An object that is tested if it has a class "subject".}

\item{term}{A subject term, for example, \code{"Data sets"}.}

\item{subjectScheme}{The name of the subject scheme or classification code or
authority if one is used. It is a namespace.}

\item{schemeURI}{The URI of the subject identifier scheme, for example
\code{"http://id.loc.gov/authorities/subjects"}}

\item{valueURI}{The URI of the subject term.
\code{"https://id.loc.gov/authorities/subjects/sh2018002256"}}

\item{prefix}{An abbreviated prefix of a scheme URI, for example,
\code{"lcch:"} representing \code{"http://id.loc.gov/authorities/subjects"}.
Widely used namespaces (schemes) have conventional abbreviations.}

\item{classificationCode}{The classificationCode subproperty may be used for
subject schemes, like ANZSRC, which do not have valueURIs for each subject term.}

\item{subject}{A subject field created by \code{\link{subject}}.}
}
\value{
A named list with the subject term, the subject scheme, URIs and prefix.

The Subject field of the dataset's attributes.
}
\description{
Create a subject
}
\examples{
subject ("Dataset",
        schemeURI = "http://id.loc.gov/authorities/subjects",
        valueURI = "https://id.loc.gov/authorities/subjects/sh2018002256",
        subjectScheme = "LCCH",
        prefix = "lcch:")

 ds <- dataset(iris,
        title = "The iris Dataset",
        author = c(
           person(family ="Anderson",
           given ="Edgar",
           role = "aut")
           ),
         identifier = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
         year = "1935",
         version = "1.0",
         description = "The famous dataset that is distributed with R.",
         url = "https://en.wikipedia.org/wiki/Iris_flower_data_set",
         resourceType = "Dataset"
         )

}
