% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_label.R
\name{var_label}
\alias{var_label}
\alias{var_label.defined}
\alias{var_label.dataset_df}
\alias{label_attribute}
\alias{var_label<-.defined}
\alias{var_label<-.dataset_df}
\title{Get / Set a variable label}
\usage{
\method{var_label}{defined}(x, ...)

\method{var_label}{dataset_df}(
  x,
  unlist = FALSE,
  null_action = c("keep", "fill", "skip", "na", "empty"),
  recurse = FALSE,
  ...
)

label_attribute(x)

\method{var_label}{defined}(x) <- value

\method{var_label}{dataset_df}(x) <- value
}
\arguments{
\item{x}{a vector or a data.frame}

\item{...}{Further potential parameters reserved for inherited classes.}

\item{unlist}{for data frames, return a named vector instead of a list}

\item{null_action}{for data frames, by default \code{NULL} will be returned for
columns with no variable label. Use \code{"fill"} to populate with the column name
instead, \code{"skip"} to remove such values from the returned list, \code{"na"} to
populate with \code{NA} or \code{"empty"} to populate with an empty string (\code{""}).}

\item{recurse}{if \code{TRUE}, will apply \code{var_label()} on packed columns
(see \code{\link[tidyr:pack]{tidyr::pack()}}) to return the variable labels of each sub-column;
otherwise, the label of the group of columns will be returned.}

\item{value}{a character string or \code{NULL} to remove the label
For data frames, with \code{var_labels()}, it could also be a named list or a
character vector of same length as the number of columns in \code{x}.}
}
\description{
See \code{\link[labelled:var_label]{labelled::var_label}} for details.
}
\keyword{internal}
