% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_labels.R
\name{var_labels}
\alias{var_labels}
\alias{var_labels.dataset}
\alias{set_var_labels.dataset}
\title{Get / Set a variable labels in a dataset}
\usage{
var_labels(x)

\method{var_labels}{dataset}(x)

\method{set_var_labels}{dataset}(x, value)
}
\arguments{
\item{x}{A dataset.}

\item{value}{A character vector for labelling the variables with
\code{set_var_labels}.}
}
\description{
Get / Set a variable labels in a dataset
}
\examples{
relabelled <- set_var_labels(
               iris_dataset,
                c(Sepal.Length="The sepal length measured in centimeters.",
                  Sepal.Width="The sepal width measured in centimeters.",
                  Species="The species of the iris observed.")
                 )
var_labels(relabelled)
}
