% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_namespace.R
\name{var_namespace}
\alias{var_namespace}
\alias{var_namespace<-}
\alias{get_variable_namespaces}
\alias{namespace_attribute}
\alias{get_namespace_attribute}
\alias{set_namespace_attribute}
\alias{namespace_attribute<-}
\title{Get / Set a namespace of measure}
\usage{
var_namespace(x, ...)

var_namespace(x) <- value

get_variable_namespaces(x, ...)

namespace_attribute(x)

get_namespace_attribute(x)

set_namespace_attribute(x, value)

namespace_attribute(x) <- value
}
\arguments{
\item{x}{a vector}

\item{...}{Further potential parameters reserved for inherited classes.}

\item{value}{a character string or \code{NULL} to remove the namespace of measure.}
}
\description{
Get / Set a namespace of measure
}
\details{
\code{get_variable_namespaces()} is identical to \code{var_namespace()}.
}
\examples{
qid = defined(c("Q275912", "Q116196078"), namespace = "https://www.wikidata.org/wiki/")
var_namespace(qid)
# To remove a namespace
var_namespace(qid) <- NULL
}
